requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
asttokens==2.4.1
billiard==4.2.1
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.63.6
django-apscheduler==0.6.2
django-crispy-forms==1.14.0
django-filter==24.1
executing==2.0.1
idna==3.7
iniconfig==2.0.0
ipython==8.22.2
jedi==0.19.1
jwt==1.3.1
kombu==5.4.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==24.2
parso==0.8.3
pip==24.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2==2.9.10
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13